+++ proto/args.proto
@@ -8,5 +8,5 @@ message Args {
   bytes adminSKI = 1;
   int64 validatorsCount = 2;
   repeated string validators = 3;
-  string CCName =4;
+  string CCName = 4;
 }
