+++ tests/unit/add_multisig_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"bytes"
