+++ tests/unit/black_list_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"github.com/anoideaopen/acl/cc/errs"
