+++ tests/unit/change_multisig_public_key_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"bytes"
