+++ tests/unit/change_public_key_with_base58_sgnature_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"crypto/ed25519"
