+++ tests/unit/get_account_info_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"encoding/json"
