+++ tests/unit/get_adresses_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"testing"
