+++ tests/unit/multisign_base58_signature_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"testing"
