+++ tests/unit/set_kyc_test.go
@@ -1,4 +1,4 @@
-package tests
+package unit
 
 import (
 	"encoding/hex"
